Fix GIS job error handling
