+++ src/main/java/com/socrata/datasync/imports2/Blueprint.java
@@ -7,10 +7,15 @@ import org.codehaus.jackson.map.annotate.JsonSerialize;
 @JsonIgnoreProperties(ignoreUnknown=true)
 @JsonSerialize(include=JsonSerialize.Inclusion.NON_NULL)
 public class Blueprint {
+    // Success fields
     private String fileId;
     private Summary summary;
-    private Error error;
     private String[] warnings;
+
+    // Error fields
+    private String code;
+    private boolean error;
+    private String message;
 
     @JsonProperty("fileId")
     public void setFileId(String fileId) {
@@ -32,13 +37,43 @@ public class Blueprint {
         return this.summary;
     }
 
+    @JsonProperty("warnings")
+    public void setWarnings(String[] warnings) {
+        this.warnings = warnings;
+    }
+
+    @JsonProperty("warnings")
+    public String[] getWarnings() {
+        return this.warnings;
+    }
+
+    @JsonProperty("code")
+    public void setCode(String code) {
+        this.code = code;
+    }
+
+    @JsonProperty("code")
+    public String getCode() {
+        return this.code;
+    }
+
     @JsonProperty("error")
-    public void setError(Error error) {
+    public void setError(boolean error) {
         this.error = error;
     }
 
     @JsonProperty("error")
-    public Error getError() {
+    public boolean getError() {
         return this.error;
+    }
+
+    @JsonProperty("message")
+    public void setMessage(String message) {
+        this.message = message;
+    }
+
+    @JsonProperty("message")
+    public String getMessage() {
+        return this.message;
     }
 }
