--- src/main/java/com/socrata/datasync/imports2/Error.java
@@ -1,21 +1,0 @@
-package com.socrata.datasync.imports2;
-
-import org.codehaus.jackson.annotate.JsonIgnoreProperties;
-import org.codehaus.jackson.annotate.JsonProperty;
-import org.codehaus.jackson.map.annotate.JsonSerialize;
-
-@JsonIgnoreProperties(ignoreUnknown=true)
-@JsonSerialize(include=JsonSerialize.Inclusion.NON_NULL)
-public class Error {
-    private String message;
-
-    @JsonProperty("message")
-    public void setMessage(String message) {
-        this.message = message;
-    }
-
-    @JsonProperty("name")
-    public String getMessage() {
-        return this.message;
-    }
-}
